CloudPebble SDK3 support, Aplite, Basalt, & Chalk
Now works for Firmware 3 (Basalt and Chalk), previously only worked for Aplite.
Removed need to configure timezone offset, this is not needed for Firmware 3 and for Aplite where this appears to be needed it is now picked up from the phone. Offset is now minutes and not hours

Update demo to use same demo value as used in https://github.com/google/google-authenticator/issues/70 - useful for test purposes.